Fixed issue #41
