+++ lib/main.js
@@ -2,30 +2,11 @@ const fs = require('fs');
 const vm = require('vm');
 const pa = require('path');
 const {EventEmitter} = require('events');
+const VMError = require('./VMError');
+const _compileToJS = require('./compileToJS');
 
 const sb = fs.readFileSync(`${__dirname}/sandbox.js`, 'utf8');
 const cf = fs.readFileSync(`${__dirname}/contextify.js`, 'utf8');
-
-const _compileToJS = function(code, compiler) {
-	if ('function' === typeof compiler) return compiler(code);
-
-	switch (compiler) {
-		case 'coffeescript':
-		case 'coffee-script':
-		case 'cs':
-		case 'text/coffeescript':
-			return require('coffee-script').compile(code, {header: false, bare: true});
-		
-		case 'javascript':
-		case 'java-script':
-		case 'js':
-		case 'text/javascript':
-			return code;
-		
-		default:
-			throw new VMError(`Unsupported compiler '${compiler}'.`);
-	}
-};
 
 /**
  * Class VM.
@@ -353,27 +334,6 @@ class NodeVM extends EventEmitter {
 		}
 
 		return new NodeVM(options).run(fs.readFileSync(filename, 'utf8'), filename);
-	}
-}
-
-/**
- * VMError.
- * 
- * @param {String} message Error message.
- * 
- * @class
- * @extends {Error}
- * @property {String} stack Call stack.
- * @property {String} message Error message.
- */
-
-class VMError extends Error {
-	constructor(message) {
-		super(message);
-		
-		this.name = 'VMError';
-
-		Error.captureStackTrace(this, this.constructor);
 	}
 }
