+++ lib/sandbox.js
@@ -2,6 +2,7 @@ const {Script} = host.require('vm');
 const fs = host.require('fs');
 const pa = host.require('path');
 const console = host.console;
+const compileToJS = host.require('./compileToJS');
 
 const BUILTIN_MODULES = host.process.binding('natives');
 const JSON_PARSE = JSON.parse;
@@ -47,7 +48,9 @@ return ((vm, host) => {
 			} else {
 				try {
 					// Load module
-					var code = `(function (exports, require, module, __filename, __dirname) { 'use strict'; ${fs.readFileSync(filename, "utf8")} \n});`;
+					var code = fs.readFileSync(filename, "utf8");
+					code = compileToJS(code, vm.options.compiler);
+					code = `(function (exports, require, module, __filename, __dirname) { 'use strict'; ${code} \n});`;
 			
 					// Precompile script
 					let script = new Script(code, {
